Handle belongs-to relation better
The belongs-to has become required by default since rails 5, 
and in general we can fix be declaring the `inverse-of`. 
However I do not want to enforce this upon the “using” classes, 
and secondly: inside vigilante it is called “operator”,
but outside the user van choose whichever “user”-class
they choose. Unfortunately declaring it here alone
does not work. 

Therefore we solve this by making the relation
“optional” which will make the nested relations 
save as expected again.